fixed critical bug (whoops)
So I thought about this yesterday but only got around to fixing it
today… I blame a lack of caffeine but I accidentally wrote in the last
version to delete the database insertions along with the deletion of
the files. WRONG.

Only deleting the files now. Must keep the entries in the database for
avoidance of duplication.

FIXED.